requirements.txt
aiogram==3.21.0
aiofiles==24.1.0
aiohttp==3.9.1
fastapi==0.104.1
python-dotenv==1.0.0
PyYAML==6.0.2
uvicorn==0.24.0
colorlog==6.9.0
coloring==0.1.8
watchfiles==1.1.0
uvloop==0.21.0
websockets==15.0.1
